feat: 참여중인 모임 목록 조회 (#142)
* feat: 참여한 모임이 있을 경우 첫 번째 모임에 참여, 아닐 경우 인트로 화면으로 이동하는 기능 구형

* feat: 참여한 모임 중 첫 번째 모임의 로그를 띄워주는 기능 구현

* refactor: 리뷰 반영

+++ android/app/src/main/java/com/woowacourse/ody/data/remote/repository/DefaultNotificationLogRepository.kt
@@ -10,6 +10,6 @@ object DefaultNotificationLogRepository : NotificationLogRepository {
     private val notificationService: NotificationService =
         RetrofitClient.retrofit.create(NotificationService::class.java)
 
-    override suspend fun fetchNotificationLogs(meetingId: Int): List<NotificationLog> =
+    override suspend fun fetchNotificationLogs(meetingId: Long): List<NotificationLog> =
         notificationService.fetchNotificationLogs(meetingId).toNotificationList()
 }
