+++ android/app/src/main/java/com/woowacourse/ody/domain/repository/NotificationLogRepository.kt
@@ -3,5 +3,5 @@ package com.woowacourse.ody.domain.repository
 import com.woowacourse.ody.domain.model.NotificationLog
 
 interface NotificationLogRepository {
-    suspend fun fetchNotificationLogs(meetingId: Int): List<NotificationLog>
+    suspend fun fetchNotificationLogs(meetingId: Long): List<NotificationLog>
 }
